lift type parser to a plugin
